rename portability-datatransfer-types ->  portability-types-transfer for consistency
