+++ portability-spi-cloud/build.gradle
@@ -16,6 +16,6 @@
 
 dependencies {
     compile ("com.fasterxml.jackson.core:jackson-databind:${jacksonVersion}")
-    compile project(':portability-data-transfer-types')
-
+    compile project(':portability-types-transfer')
+                                                       
 }
