+++ portability-spi-transfer/build.gradle
@@ -16,6 +16,6 @@
 
 dependencies {
 
-    compile project(':portability-data-transfer-types')
+    compile project(':portability-types-transfer')
 
 }
