+++ portability-types-transfer/src/main/java/org/dataportabilityproject/types/transfer/EntityType.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.datatransfer.types;
+package org.dataportabilityproject.types.transfer;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
