+++ portability-types-transfer/src/main/java/org/dataportabilityproject/types/transfer/PortableType.java
@@ -13,12 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.datatransfer.types.auth;
+package org.dataportabilityproject.types.transfer;
 
-import org.dataportabilityproject.datatransfer.types.PortableType;
+import com.fasterxml.jackson.annotation.JsonTypeInfo;
 
 /**
- * Base type for authentication data.
+ * Base model type that supports language interoperability and extensibility.
+ *
+ * Subtypes use {@link com.fasterxml.jackson.annotation.JsonTypeName} to define the concrete type key for de/serialization.
  */
-public abstract class AuthData extends PortableType {
+@JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.PROPERTY, property = "ptype")
+public abstract class PortableType {
 }
