+++ portability-types-transfer/src/main/java/org/dataportabilityproject/types/transfer/package-info.java
@@ -16,5 +16,5 @@
 /**
  * Defines the data portability model objects.  Model objects represent the flow of data through the system.
  */
-package org.dataportabilityproject.datatransfer.types;
+package org.dataportabilityproject.types.transfer;
