+++ portability-types-transfer/src/test/java/org/dataportabilityproject/transfer/types/job/PortabilityJobTest.java
@@ -1,7 +1,8 @@
-package org.dataportabilityproject.datatransfer.types.job;
+package org.dataportabilityproject.transfer.types.job;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.dataportabilityproject.test.types.ObjectMapperFactory;
+import org.dataportabilityproject.types.transfer.job.PortabilityJob;
 import org.junit.Assert;
 import org.junit.Test;
